chore: added a generic zodParse function
